Factor out the template disallowed functions into its own package so it can be referenced from multiple higher level packages
