+++ server/graphdb/internal/services/template.go
@@ -24,7 +24,7 @@ import (
 	"log/slog"
 	"text/template"
 
-	"github.com/Masterminds/sprig/v3"
+	"github.com/specterops/bloodhound/packages/go/safetemplate"
 )
 
 // NodeContext is the template representation of a node.
@@ -64,38 +64,6 @@ type kindInfoContent struct {
 	} `json:"markdown"`
 }
 
-// we don't support these template functions
-var unsupportedFns = []string{
-	"bcrypt",
-	"htpasswd",
-	"genPrivateKey",
-	"derivePassword",
-	"buildCustomCert",
-	"genCA",
-	"genCAWithKey",
-	"genSelfSignedCert",
-	"genSignedCert",
-	"genSignedCertWithKey",
-	"encryptAES",
-	"decryptAES",
-	"regexMatch",
-	"regexFindAll",
-	"regexFind",
-	"regexReplaceAll",
-	"regexReplaceAllLiteral",
-	"regexSplit",
-	"mustRegexMatch",
-	"mustRegexFindAll",
-	"mustRegexFind",
-	"mustRegexReplaceAll",
-	"mustRegexReplaceAllLiteral",
-	"mustRegexSplit",
-	"urlParse",
-	"urlJoin",
-	"randint",
-	"ago",
-}
-
 func renderKindInfoMarkdown(content json.RawMessage, context any) (string, error) {
 	var contentView kindInfoContent
 
@@ -103,13 +71,8 @@ func renderKindInfoMarkdown(content json.RawMessage, context any) (string, error
 		return "", fmt.Errorf("unmarshalling markdown content: %w", err)
 	}
 
-	functions := sprig.HermeticTxtFuncMap()
-	for _, functionName := range unsupportedFns {
-		delete(functions, functionName)
-	}
-
 	parsedTemplate, err := template.New("kind-info-markdown").
-		Funcs(functions).
+		Funcs(safetemplate.FuncMap()).
 		Parse(contentView.Markdown.Content)
 	if err != nil {
 		return contentView.Markdown.Content, fmt.Errorf("parsing markdown template: %w", err)
